LIM-255, LIM-64, LIM-63
Fix date of birth issue and UI fixes

+++ app/src/main/kotlin/io/square1/limor/scenes/main/viewmodels/UpdateUserViewModel.kt
@@ -27,7 +27,7 @@ class UpdateUserViewModel @Inject constructor(private val profileUseCase: Profil
     var description: String = ""
     var email: String = ""
     var phone_number: String = ""
-    var date_of_birth: Int = 0
+    var date_of_birth: Long = 0
     var gender: String = ""
     var image = ""
     var notifications_enabled = false
